Updating all programs & libraries in initrd.img and install.img
Remove & replace all the old stuff, we will track packages from Debian 13
as I think that it's the best distro to get i386/i686 packages nowdays.
Also I made a README file to list all the packages we need to track. If
anyone want to make an automated script to update, feel free to do it.

Also, busybox while still on the same version number, will use Debian version
and it's non-static so it gives faster speed (no wonder why cwhuang use it).

Lastly, there is one case that will use the program inside the system.img instead
of install.img and that is mkntfs (or mkfs.ntfs). We are now building mkfs.ntfs from
ntfs-3g, switching to it is fine from what I tested.

Goodbye TrustedGrub.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>