init: Fix disk detection bug when choosing install
sometimes check_root() might accidentally scan internal drive, making
the init script failed to extract install.img or just looping in that
while :; do loop

So if we boot with INSTALL=1 (which start the installer), we will assume
that install.img and another dir in the iso exist (here I choose isolinux).
If they're not exist, stop the check_root() function to skip to other drive.

We also put some small modification here that are not related to this fix.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>