Use rsync to copy files and calculate the percentage
Previous method of copying the required files to disk is to use cpio & pv.
However, this method is started to show some of it's limitation, there is
one report that busybox cpio is showing some weird bug, not to mention the
progress is now reach 200%, which is not a pleasant look.

I noticed that Bliss, like a lot of custom ROMs, get rsync source from LineageOS
and build it. So instead of trying to fix the cpio command to copy files, I switched
to rsync. Although the current command will need to use the GNU version of sed because
of --unbuffered and also stdbuf, it seems to create a stable copying process, not to
mention the percentage now looks better.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>